Move middleware into /lib directory
The /app/middleware dir is automatically added to the list of autoload paths in the parent app. It doesn't really make sense for middleware to ever be hot reloaded, as the middleware stack setup does not get re-run. Any changes to the middleware would not have any effect.
